extract dom source mapping transformation
